+++ Cookie.API/Protocol/Network/Types/Game/Context/MapCoordinatesAndId.cs
@@ -6,7 +6,7 @@ namespace Cookie.API.Protocol.Network.Types.Game.Context
     {
         public new const ushort ProtocolId = 392;
 
-        public MapCoordinatesAndId(int mapId)
+        public MapCoordinatesAndId(double mapId)
         {
             MapId = mapId;
         }
@@ -16,18 +16,18 @@ namespace Cookie.API.Protocol.Network.Types.Game.Context
         }
 
         public override ushort TypeID => ProtocolId;
-        public int MapId { get; set; }
+        public double MapId { get; set; }
 
         public override void Serialize(IDataWriter writer)
         {
             base.Serialize(writer);
-            writer.WriteInt(MapId);
+            writer.WriteDouble(MapId);
         }
 
         public override void Deserialize(IDataReader reader)
         {
             base.Deserialize(reader);
-            MapId = reader.ReadInt();
+            MapId = reader.ReadDouble();
         }
     }
 }
